backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
reportlab==4.1.0
openpyxl==3.1.2
requests==2.31.0
Werkzeug==3.0.1
